requirements.txt
Django>=3.1.3, <3.2.0
djangorestframework>=3.12.1, <3.13.0
psycopg2>=2.8.6,<2.9.0
Pillow >=5.3.0,<5.4.0

flake8>=3.6.0,<3.7.0 